Made Aurora work with GFX interface.
Yes, it's ironice to get Aurora written for SmartMatrix, work with
SmartMatrix again through the GFX interface, but it's a good test.

Added drawPixel(int16_t x, int16_t y, CRGB color) to make life simpler.

+++ examples/Aurora/Effects.h
@@ -57,7 +57,7 @@ void drawForegroundHLine(int16_t x0, int16_t x1, int16_t y) {
   int i;
 
   for (i = x0; i <= x1; i++) {
-    matrix->drawPixel(i, y, 1);
+    matrix->drawPixel(i, y, (uint32_t)1);
   }
 }
